dovie_messenger_python/requirements.txt
# FastAPI and ASGI server
fastapi==0.104.1
uvicorn[standard]==0.24.0

# Database
sqlalchemy==2.0.23
alembic==1.13.0
asyncpg==0.29.0
psycopg2-binary==2.9.9

# Authentication and Security
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.3.0
python-multipart==0.0.6
bcrypt==4.1.2

# WebSocket support
websockets==12.0

# HTTP client for external APIs
httpx==0.25.2
aiofiles==23.2.1

# Push notifications
pywebpush==1.14.0
cryptography==41.0.8

# SMS/Phone verification (Twilio)
twilio==8.10.2

# Image processing
pillow==10.1.0

# QR codes
qrcode[pil]==7.4.2

# OpenAI API
openai==1.3.5

# Environment variables
python-dotenv==1.0.0

# Date/time utilities
python-dateutil==2.8.2

# Validation
pydantic==2.5.0
email-validator==2.1.0

# File handling
python-magic==0.4.27

# Apple Push Notifications (iOS)
apns2==0.7.2

# Session management
redis==5.0.1

# Background tasks
celery==5.3.4

# CORS support
fastapi-cors==0.0.6

# Static files
jinja2==3.1.2

# Logging
structlog==23.2.0